refactor: Move parser and lexer tests into integration tests.
Parser and lexer tests are too large for unit tests and are
rather integration tests.
Therefore, they are now in the tests directory.
Furthermore, use test_common for the salsa test databases.
Must be in its own crate because for some reason cannot find
the files otherwise if they are in the tests directory.