refactor: Use NonEmptyVec for Trisult errors
Trisult always had the requirement that the errors vec can never be null.
However, this was not properly checked.
Furthermore, NonEmptyVec can store one
element inline.
When merging two NonEmptyVec, it will try merge errors while preventing reallocation of the underlying vec. As a tradeoff order of
elements after merging is not preserved.